requirements.txt
pandas
requests
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
python-dotenv